refactor: changed architecture to use indexing
