+++ src/components/atoms/Div/Div.tsx
@@ -1,8 +1,8 @@
 import styled from '@xstyled/styled-components'
 
-const Container = styled.divBox`
+const Div = styled.divBox`
   display: flex;
   width: fit-content;
 `
 
-export default Container
+export default Div
