+++ src/components/atoms/Div/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Div'
