--- src/components/molecules/Search/index.tsx
@@ -1,14 +1,0 @@
-import Icon from 'components/atoms/Icon'
-
-import { Input, Container, IconContainer } from './style'
-
-const Search = ({ ...props }) => (
-  <Container {...props}>
-    <IconContainer>
-      <Icon type='search' />
-    </IconContainer>
-    <Input type='search' />
-  </Container>
-)
-
-export default Search
